Fix downcast warnings
Add explicit casting to some assignments that have warnings due to implicit casting to smaller types.

+++ src/curve25519/ed25519/nacl_sha512/hash.c
@@ -26,7 +26,7 @@ int crypto_hash_sha512(unsigned char *out,const unsigned char *in,unsigned long 
 {
   unsigned char h[64];
   unsigned char padded[256];
-  int i;
+  unsigned long long i;
   unsigned long long bytes = inlen;
 
   for (i = 0;i < 64;++i) h[i] = iv[i];
